Avoid reference count manipulation iterating over parent Contexts
These loops iterate from a `Context`'s `ContextSharedState` (to which the `Context` holds a reference) up the chain of `parent` pointers (again, to which references are held). Since nothing in the chain may change while we are iterating over it, we need not hold additional references to the elements we're examining - it's sufficient to use a raw pointer for the iteration variable.