Add Storage Performance tests for DownloadTo API (#2987)
Adding 3 new tests for Storage Performance
The next tests are expected to find out the overhead between each layer of the Azure SDK client while performing a download operation. The current performance test is measuring the client + azure core pipeline + http transport.

The next tests removes authentication policy and then each layer up to the HTTP transport adapter directly.

DownloadBlobSas
Generates a SaS token to download a blob. This way the authentication policy "sharedKey" won't run in the pipeline.

DownloadBlobWithPipelineOnly
Use the SaS token as well but send the request without the BlobClient. Instead it uses the Azure Core Pipeline directly

DownloadBlobWithTransportOnly
Same as before, the SaS token is used to download the blob but consuming only the libcurl-transport adapter. This option skips the HttpPipeline. The only elements used from Azure Core are the HttpRequest and the Azure::Context

Note
porting tests from the .NET equivalent:

https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/storage/Azure.Storage.Blobs/perf/Azure.Storage.Blobs.Perf/Scenarios/DownloadSasUriBlobClient.cs

https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/storage/Azure.Storage.Blobs/perf/Azure.Storage.Blobs.Perf/Scenarios/DownloadSasUriHttpClient.cs

https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/storage/Azure.Storage.Blobs/perf/Azure.Storage.Blobs.Perf/Scenarios/DownloadSasUriHttpPipeline.cs